requirements.txt
transformers
aiofiles==24.1.0
aiohttp==3.9.5
alembic==1.13.1
cachetools==5.3.3
cryptography==42.0.5
fastapi==0.115.0
httpx==0.27.2
huggingface_hub==0.23.4
jsonschema==4.23.0
ngrok==1.4.0
numpy
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-multipart==0.0.12
packaging==24.1
passlib==1.7.4
Pillow==10.4.0
prometheus_client==0.20.0
py_localtunnel==1.0.3
PyJWT==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
uvicorn==0.31.0
vllm==0.6.2
